Apply Garrett's product code CR comments:
core/lib/util/percentage-scaler:
Added a new type to implement percentage of the entire archive in addition to percentage of the individual file entry in the zip.

archive.ts:
Add percentage of the entire archive in terms of the above, plus other nitpicks.

archive-tests:
Fixed a bug in the test where we were catching the test failure exception.

streams.ts:
Fixed a bug where the throw is breaking node and emit an error event instead; use percentage-scaler.