Implement persistence of modification time from zips and fix subtle EnhancedReadable bug.
Node readable streams have an odd behavior where they have side effects triggered by attaching a handler for the 'data' event; the EnhancedReadable design we were previously using attached such a handler in order to emit progress events. This caused subtle async bugs where sometimes the content from the readable stream was dropped on the floor or sometimes triggered hangs.

The change removes the "enhanceXxx" stream machinery, and instead resolves the Promise/streams impedence mismatch using pipeline from Node's streams module.

Additional changes:

local-filesystem.ts:
Use open and futimes in order to set the modification time when creating a file if selected.

acquire-tests.ts:
Remove tests for emission of the progress event we no longer emit. It is likely we'll want to restore progress reporting in some form but this requires a change in interface.

archive-tests.ts:
Test that the modified time is correct for one of the files in one of the tests. Disable the FailsToUnpackMalformed test because the workarounds attempted before relied on the now-known-to-be-problematic data event handler. I tested with a local / vendored copy of yazul with https://github.com/thejoshwolfe/yauzl/pull/123 applied and that makes the test pass; if we experience problems with hangs in the future we might want to consider vendoring yauzl.