Follow the vcpkg Z_ convention for internal things.
vcpkg follows a convention for user-exposed things like switches and environment variables:

plain things are user settings with full support intended to be stabilized "forever"
X_ things are "experimental" things which eventually are intended to enter the previous "full support" category.
Z_ things are "internal, not intended for users to touch" and can change at any time.

As the variables VCPKG_POSTSCRIPT and VCPKG_UNDO are in the "internal" bucket, they should get Z_.